pyproject.toml
[tool.poetry]
name = "playlistgen"
version = "0.1.0"
description = ""
authors = ["Harold Abnabit <_@habnab.it>"]

[tool.poetry.dependencies]
python = "^3.5"
attrs = "^18.1"
click = "^6.7"
numpy = "^1.15"
py-applescript = "^1.0"
pyobjc-framework-applescriptkit = "^4.2"
tqdm = "^4.24"
pyramid = "^1.10"
waitress = "^1.3"
arrow = "^0.14.2"
cornice = "^3.5"
marshmallow = {version = "^2.19",allows-prereleases = true}
parsimonious = "^0.8.1"

[tool.poetry.scripts]
playlistgen = "playlistgen.playlistgen:main"

[tool.poetry.dev-dependencies]
plop = "^0.4.0"
flameprof = "^0.4.0"
